capn-rpc-hello.py/pyproject.toml
[project]
name = "capn-rpc-hello-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pycapnp>=2.2.1",
]
